refactor: Tidy up interfaces and add a team service
