+++ src/services/team/index.ts
@@ -1,0 +1,5 @@
+import { apiRequest } from "@/api";
+import { Team, Profile } from "./interfaces";
+
+export const getTeam = (): Promise<Team> => apiRequest<Team>("/team/", "GET");
+export type { Profile, Team };
