+++ src/wrappers/ManageWrapper.tsx
@@ -6,8 +6,12 @@ import {
   ProvideAuth /* TODO export Services */,
 } from "@gliff-ai/manage";
 import { DominateStore, API_URL } from "@/store";
-import { useAuth, UserAccess } from "@/hooks/use-auth";
-import { inviteNewCollaborator, inviteNewUser } from "@/services/user";
+import { useAuth } from "@/hooks/use-auth";
+import {
+  inviteNewCollaborator,
+  inviteNewUser,
+  UserAccess,
+} from "@/services/user";
 import { trustedServicesAPI, TrustedService } from "@/services/trustedServices";
 import { jsPluginsAPI, JsPlugin } from "@/services/plugins";
 import { GalleryTile } from "@/interfaces";
